Make color cube app write to logcat on Android (#68)
* Make color cube app write to logcat on Android

The color cube test app does all its error reporting and logging
using std::cout. This works fine on desktop platforms but not
so well on Android. Android apps typically write to logcat, as
the standard handles are not easily accessible.

This introduces a AMDVulkanDemoVkUtils::Log() which writes to
logcat on Android, and stdout everywhere else. This also
changes all std::cout writes to use the new utility method.

Finally, the code that tries to create a Vulkan android surface
was neglecting to check if the operation succeeded. It wrote the
result code to stdout, but assumed success. It now checks and
aborts if the creation failed.

Change-Id: I9c731a3363227e3fc14289445f112526e1882689

* Avoid memory leak in error case