Init the target buffer of vsnprintf to empty string
In the scenario where vsnprintf() fails and doesn't write anything
to the target buffer, we are sending the buffer forward for logging.
The buffer has uninitialized memory at that point. Best case scenario:
we print garbage. Worst case scenario, we do a read overrun and
crash (unlikely since it's stack memory).