Various improvements to GPA's logging facility (#71)
* Various improvements to GPA's logging facility

1. LogDebugXxxxx() methods are variadic printf based functions, but
LogXxxxx() methods are not. Upgraded the latter for consistency
and to make it easier to log dynamic non-debug messages

2. The LogDebugXxxxx() methods could trash memory. Should use
vsnprintf instead of vsprintf.

3. The LogDebugXxxxx() methods had a lot of redundant/common
logic.

* Additional logging tweaks

When building for host, the need or additional log statement
adjustments surfaced.

Change-Id: I2855ade49a2e798c046986ad64661eeac54a430d

* Improved logging of device ID

Change-Id: Ibd51cd822ef47473c71e85c6f97daddb1cdae59f

* Init the target buffer of vsnprintf to empty string

In the scenario where vsnprintf() fails and doesn't write anything
to the target buffer, we are sending the buffer forward for logging.
The buffer has uninitialized memory at that point. Best case scenario:
we print garbage. Worst case scenario, we do a read overrun and
crash (unlikely since it's stack memory).