The VkColorCube app doesn't log failure status
The failure status code being returned via the GPA calls
aren't being logged by the VkColorCube app. It does log that
what it's trying to do failed, but not why it failed (according
to the GPA implementation). In my case a C++ exception was
being thrown and that fact was just not surfaced at all.

Also changed the textual string for kGpaStatusErrorException
to mention it's specifically a C++ exception, as otherwise
the word "exception" is generic/vague and could be interepreted
as such.